refactor: visitLog에서 entrance로 패키지명 변경 및 swagger 수정
- 퇴장 API가 불필요해짐에 따른 패키지명 변경
- 방문자 기록임을 명시하도록 swagger tag 변경
