+++ src/main/java/com/lgcns/domain/entrance/domain/UserGender.java
@@ -1,4 +1,4 @@
-package com.lgcns.domain.visitLog.domain;
+package com.lgcns.domain.entrance.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
