+++ src/main/java/com/lgcns/domain/entrance/repository/EntranceRepository.java
@@ -1,6 +1,6 @@
-package com.lgcns.domain.visitLog.repository;
+package com.lgcns.domain.entrance.repository;
 
-import com.lgcns.domain.visitLog.domain.Entrance;
+import com.lgcns.domain.entrance.domain.Entrance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EntranceRepository extends JpaRepository<Entrance, Long> {}
