+++ src/main/java/com/lgcns/domain/entrance/service/EntranceService.java
@@ -1,0 +1,7 @@
+package com.lgcns.domain.entrance.service;
+
+import com.lgcns.domain.entrance.dto.request.EntranceCreateRequest;
+
+public interface EntranceService {
+    void createEntrance(EntranceCreateRequest request);
+}
