--- src/main/java/com/lgcns/domain/visitLog/service/VisitLogService.java
@@ -1,7 +1,0 @@
-package com.lgcns.domain.visitLog.service;
-
-import com.lgcns.domain.visitLog.dto.request.EntranceCreateRequest;
-
-public interface VisitLogService {
-    void createEntrance(EntranceCreateRequest request);
-}
